Add Mac support to the SConscript, and fix some compile warnings
git-svn-id: https://rpg.hamsterrepublic.com/source/wip@4927 7d344553-34f0-0310-a9b1-970ce8f1c3a2

+++ filelayer.cpp
@@ -98,7 +98,7 @@ static FB_FILE_HOOKS lumpfile_hooks = {
 };
 
 void dump_openfiles() {
-	debug(2, "%d open files:", openfiles.size());
+	debug(2, "%d open files:", (int)openfiles.size());
 	for (openfiles_iterator_t it = openfiles.begin(); it != openfiles.end(); ++it) {
 		const char *fname = it->second.name.c_str();
 		debug(2, " %p (%s)", it->first, fname);
